move correct events to be every chunk
