+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/jfr/events/EveryChunkNativePeriodicEvents.java
@@ -28,13 +28,19 @@ import java.lang.management.ManagementFactory;
 import java.lang.management.ThreadMXBean;
 
 import org.graalvm.nativeimage.StackValue;
+import org.graalvm.nativeimage.IsolateThread;
 
 import com.oracle.svm.core.Uninterruptible;
+import com.oracle.svm.core.heap.Heap;
+import com.oracle.svm.core.heap.PhysicalMemory;
+import com.oracle.svm.core.heap.VMOperationInfos;
 import com.oracle.svm.core.jfr.JfrEvent;
 import com.oracle.svm.core.jfr.JfrNativeEventWriter;
 import com.oracle.svm.core.jfr.JfrNativeEventWriterData;
 import com.oracle.svm.core.jfr.JfrNativeEventWriterDataAccess;
 import com.oracle.svm.core.jfr.JfrTicks;
+import com.oracle.svm.core.thread.VMThreads;
+import com.oracle.svm.core.thread.JavaVMOperation;
 
 import jdk.jfr.Event;
 import jdk.jfr.Name;
@@ -49,7 +55,9 @@ public class EveryChunkNativePeriodicEvents extends Event {
         emitJavaThreadStats(threadMXBean.getThreadCount(), threadMXBean.getDaemonThreadCount(),
                         threadMXBean.getTotalStartedThreadCount(), threadMXBean.getPeakThreadCount());
 
-        emitPhysicalMemory(com.oracle.svm.core.heap.PhysicalMemory.size().rawValue(), 0);
+        emitPhysicalMemory(PhysicalMemory.size().rawValue(), 0);
+        emitClassLoadingStatistics(Heap.getHeap().getClassCount());
+        emitPerThreadEveryChunkEvents();
     }
 
     @Uninterruptible(reason = "Accesses a JFR buffer.")
@@ -79,6 +87,51 @@ public class EveryChunkNativePeriodicEvents extends Event {
             JfrNativeEventWriter.putLong(data, totalSize);
             JfrNativeEventWriter.putLong(data, usedSize);
             JfrNativeEventWriter.endSmallEvent(data);
+        }
+    }
+
+    @Uninterruptible(reason = "Accesses a JFR buffer.")
+    private static void emitClassLoadingStatistics(long loadedClassCount) {
+        if (JfrEvent.ClassLoadingStatistics.shouldEmit()) {
+            JfrNativeEventWriterData data = StackValue.get(JfrNativeEventWriterData.class);
+            JfrNativeEventWriterDataAccess.initializeThreadLocalNativeBuffer(data);
+
+            JfrNativeEventWriter.beginSmallEvent(data, JfrEvent.ClassLoadingStatistics);
+            JfrNativeEventWriter.putLong(data, JfrTicks.elapsedTicks());
+            JfrNativeEventWriter.putLong(data, loadedClassCount);
+            JfrNativeEventWriter.putLong(data, 0); /* unloadedClassCount */
+            JfrNativeEventWriter.endSmallEvent(data);
+        }
+    }
+
+    /**
+     * This is responsible for handling the emission of all per thread events. This allows for a
+     * single VM operation to iterate the running threads.
+     */
+    private static void emitPerThreadEveryChunkEvents() {
+        if (shouldEmitUnsafe()) {
+            EmitPerThreadEveryChunkEventsOperation vmOp = new EmitPerThreadEveryChunkEventsOperation();
+            vmOp.enqueue();
+        }
+    }
+
+    @Uninterruptible(reason = "Used to avoid the VM operation if it is not absolutely needed.")
+    private static boolean shouldEmitUnsafe() {
+        /* The returned value is racy. */
+        return JfrEvent.ThreadCPULoad.shouldEmit() || JfrEvent.ThreadAllocationStatistics.shouldEmit();
+    }
+
+    private static final class EmitPerThreadEveryChunkEventsOperation extends JavaVMOperation {
+        EmitPerThreadEveryChunkEventsOperation() {
+            super(VMOperationInfos.get(EmitPerThreadEveryChunkEventsOperation.class, "Emit per thread every chunk events", SystemEffect.SAFEPOINT));
+        }
+
+        @Override
+        protected void operate() {
+            for (IsolateThread isolateThread = VMThreads.firstThread(); isolateThread.isNonNull(); isolateThread = VMThreads.nextThread(isolateThread)) {
+                ThreadCPULoadEvent.emit(isolateThread);
+                ThreadAllocationStatisticsEvent.emit(isolateThread);
+            }
         }
     }
 }
